requirements.txt
# Data Processing & Database
pandas==2.2.2
duckdb==0.10.0
pyarrow==15.0.0

# Backend API
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Frontend & Visualization
streamlit==1.31.0
plotly==5.18.0
matplotlib==3.8.2

# Utilities
pyyaml==6.0.1